Get simple sample to run.
